Now uses "Trust Region Reflective" least_squares
This is considered more robust that the previous Levenberg-Marquardt
algorithm. (I found a specific problem with a linear fit y = a x + b
where a was exactly zero in the fit result. That led to infinite
covariances with L-M.)

Improved printing of near-zero numbers and of numbers with infinite
uncertainty.